% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bar_stat_units}
\alias{bar_stat_units}
\title{Similarities and differences of mRNAs co-expressed with lncRNAs between two experimental conditions.}
\usage{
bar_stat_units(object.list)
}
\arguments{
\item{object.list}{A list of objects including two scLNC objects that need to be compared.}
}
\value{
A graph with combined bar and dot plots.
}
\description{
Similarities and differences of mRNAs co-expressed with lncRNAs between two experimental conditions.
}
