% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{CaculateCorelation}
\alias{CaculateCorelation}
\title{Calculate the co-expression correlation.}
\usage{
CaculateCorelation(object, fileName)
}
\arguments{
\item{object}{A scLNC object including expression matrix slot.}

\item{fileName}{The name of the output file.}
}
\value{
Return a co-expression correlation matrix and a co-expression correlation list including genes, co-expression relationship pairs, gene types and locus biotypes.
}
\description{
Calculate the co-expression correlation.
}
\examples{
\dontrun{
library('bigSCale')
data(LNCobject)
data(scLNCgencode)
rdata=CaculateCorelation(object=LNCobject,fileName='HCC')
}
#'
}
