% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{celltype_enrich}
\alias{celltype_enrich}
\title{Identify the enrichment of cell types under different experimental conditions.}
\usage{
celltype_enrich(object = LRAT.object)
}
\arguments{
\item{object}{A scLNC object.}
}
\value{
A heatmap of cell type enrichment.
}
\description{
Identify the enrichment of cell types under different experimental conditions.
}
