% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{classification}
\alias{classification}
\title{Pie chart of gene type statistics.}
\usage{
classification(object, genetype_range = "all")
}
\arguments{
\item{object}{A scLNC object.}

\item{genetype_range}{The range of gene types, must be one of "all" (default), "lncRNA" or "lncRNA_overlap".}
}
\value{
A plot of gene type statistics
}
\description{
Pie chart of gene type statistics.
}
