% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{CompareGo}
\alias{CompareGo}
\title{Visualization of functional enrichment results of metascape multiple gene lists.}
\usage{
CompareGo(gores, cutoff_p = 0.05, cutoff_top = 10)
}
\arguments{
\item{gores}{Metascape results.}

\item{cutoff_p}{P-value cut off. Default is 0.05.}

\item{cutoff_top}{Top markers cut off.}
}
\value{
Return a dotplot.
}
\description{
Visualization of functional enrichment results of metascape multiple gene lists.
}
\examples{
\dontrun{
data(MetascapeGoMultG)
CompareGo(gores=MetascapeGoMultG,cutoff_p = 0.05,cutoff_top = 10)
}

}
