% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CorDensity}
\alias{CorDensity}
\title{Probability density plot of the paired co-expression correlation coefficients of lncRNAs and their nearby mRNAs.}
\usage{
CorDensity(CoexpPairs)
}
\arguments{
\item{CoexpPairs}{A data.frame where the columns are correlation coefficient and distance of lncRNAs and their nearby genes.}
}
\value{
A probability density plot.
}
\description{
Probability density plot of the paired co-expression correlation coefficients of lncRNAs and their nearby mRNAs.
}
