% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{createLRAT}
\alias{createLRAT}
\title{Create a scLNC object}
\usage{
createLRAT(
  count_table,
  cell_info,
  min.cells = 0,
  min.genes = 0,
  proj.name = "test",
  gtf = scLNCgencode
)
}
\arguments{
\item{count_table}{Gene count matrix.}

\item{cell_info}{A list of cell information.}

\item{min.cells}{Include features detected in at least this many cells. Default is 0.}

\item{min.genes}{Include cells where at least this many genes are detected. Default is 0.}

\item{proj.name}{Project name.}

\item{gtf}{Gene annotation from Gencode. Default is Gencode.v32.}
}
\value{
Return a scLNC object with genes and cells information.
}
\description{
Create a scLNC object from a gene expression matrix and a list of cell information.
The expected format of the input matrix is genes x cells and cells x item, respectively.
}
\examples{
\dontrun{
data(HCC_rawcount)
data(HCC_cellinfo)
data(scLNCgencode)
LNCobject <- createLRAT(count_table = HCC_rawcount, cell_info = HCC_cellinfo,
 min.cells = 15, min.genes = 200, proj.name = "HCC")
}

}
