% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{DeActivity}
\alias{DeActivity}
\title{Differential activity of units between two groups.}
\usage{
DeActivity(object, item, FC = 0.1, pvalue = 0.05, min.pct = 0.3, padj = 1, ...)
}
\arguments{
\item{object}{A scLNC object.}

\item{item}{Attribute for comparison.}

\item{FC}{Limit testing to units which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.1.}

\item{pvalue}{P-value cut off. Default is 0.05.}

\item{min.pct}{Only test units that are obtained activity score in a minimum fraction of min.pct cells in either of the two populations. Default is 0.3.}

\item{padj}{P-value adjust cut off. Default is 1.}

\item{...}{Other options used to control matching behavior between duplicate strings. Passed on to \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}}.}
}
\value{
Return a scLNC object with Differential activity of units slot.
}
\description{
Differential activity of units between two groups.
}
\examples{
\dontrun{
data(LNCobject2)
LNCobject <- AUCell_score(object = LNCobject2,lnclist = NULL,min.target = 0)
LNCobject<- DeActivity(object=LNCobject,item='majorCluster',
FC=0.1,pvalue=0.05,min.pct=0.3,padj=1)
}
}
