% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{DEAUC_2item}
\alias{DEAUC_2item}
\title{Title The difference in AUC between the two experimental conditions.}
\usage{
DEAUC_2item(
  object.list,
  cluster_a = 5,
  cluster_b = 5,
  limit_a = 50,
  limit_b = 50,
  label_a = "group1",
  label_b = "group2"
)
}
\arguments{
\item{object.list}{A list of objects including two scLNC objects that need to be compared.}

\item{cluster_a}{The number of divisions of clusters in a group.}

\item{cluster_b}{The number of divisions of clusters in the other group.}

\item{limit_a}{The number of lncRNAs to display in a group.}

\item{limit_b}{The number of lncRNAs to display in the other group.}

\item{label_a}{The title of a group.}

\item{label_b}{The title of the other group.}
}
\value{
The AUC heatmaps of two groups and the same lncRNAs in the two groups.
}
\description{
Title The difference in AUC between the two experimental conditions.
}
