% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{display_unit}
\alias{display_unit}
\title{The sames and differences of mRNAs coexpressed with lncRNA in an unit between two groups.}
\usage{
display_unit(
  object1,
  object2,
  myunit = "LINC010273",
  corcut1 = 0.7,
  corcut2 = 0.5
)
}
\arguments{
\item{object1}{A scLNC object of group1.}

\item{object2}{A scLNC object of group2.}

\item{myunit}{Interesting lncRNA unit name.}

\item{corcut1}{Top pairs with high correlation coefficient in group1.}

\item{corcut2}{Top pairs with high correlation coefficient in group2.}
}
\value{
Return a network diagram.
}
\description{
The network diagram of units between the two groups included the same lncRNA and differences mRNAs co-expressed with them.
}
\examples{
\dontrun{
data(LNCobject2)
display_unit(object1=LNCobject2,object2=LNCobject2,
myunit='LINC00861',corcut1=0.7,corcut2=0.5)
}

}
