% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DotFeatures}
\alias{DotFeatures}
\title{Dotplot of gene expression.}
\usage{
DotFeatures(
  object = LRAT.object,
  features = unique(DE_l$gene),
  item = "",
  mytitle = NULL,
  split.by = NULL,
  mincell.peritem = 15
)
}
\arguments{
\item{object}{A scLNC object.}

\item{features}{Interesting gene ensemble ids, such as the calculated differentially expressed lncRNAs.}

\item{item}{A variable to split the dotplots by.}

\item{mytitle}{The title of dotplot.}

\item{split.by}{A experimental condition variables to group cells by.}

\item{mincell.peritem}{Minimum number of cells per cell type.}
}
\value{
A dotplot.
}
\description{
Dotplot of gene expression.
}
