% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{draw_shortPath}
\alias{draw_shortPath}
\title{Draw shortPath plot}
\usage{
draw_shortPath(corrlist_shortPath)
}
\arguments{
\item{corrlist_shortPath}{A co-expression list between lncRNAs and target genes, including both lncRNA-mRNA and mRNA-mRNA pairs.}
}
\description{
Draw shortPath plot
}
