% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{FilterPairs}
\alias{FilterPairs}
\title{Filtering of co-expressed relationship pairs.}
\usage{
FilterPairs(data, corcut = 0.6, RPSL = FALSE, targetcut = 0, object)
}
\arguments{
\item{data}{A co-expressed correlation list including genes, co-expressed relationship pairs and gene types.}

\item{corcut}{Choose the relationship pairs with higher correlation coefficient.}

\item{RPSL}{Boolean values determining if co-expressed relationship pairs should be removed the ribosome gene. Default is FALSE.}

\item{targetcut}{Selected lncRNAs with higher number of co-expressed mRNA. Default is 0.}

\item{object}{A scLNC object.}
}
\value{
Return a scLNC object with co-expressed relationship pairs slot.
}
\description{
Filtering of co-expressed relationship pairs.
}
\examples{
\dontrun{
data(CoExpreList)
data(LNCobject)
LNCobject=FilterPairs(data=CoExpreList, object=LNCobject, corcut=0.6,
RPSL=FALSE, targetcut=0)
}

}
