% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{getIndirectNodes}
\alias{getIndirectNodes}
\title{Identify the indirect correlated mRNAs with input lncRNA.}
\usage{
getIndirectNodes(corrlist_shortPath)
}
\arguments{
\item{corrlist_shortPath}{A co-expression list between lncRNAs and target genes, including both lncRNA-mRNA and mRNA-mRNA pairs.}
}
\value{
A vector of non-directly correlated mRNAs for lncRNA.
}
\description{
Identify the indirect correlated mRNAs with input lncRNA.
}
\examples{
\dontrun{
library(dplyr)
data(CoExpreList_lm_mm)
ID = 'ENSG00000204054'
end = unique((corrlist \%>\% dplyr::filter(simple_type.x == 'lncRNA',
simple_type.y == 'mRNA', row == lncID))$column)
corrlist_shortPath = getShortPath(lncID=ID,
endpoints=end, corrlist=CoExpreList_lm_mm)
getIndirectNodes(corrlist_shortPath)
}

}
