% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{getShortPath}
\alias{getShortPath}
\title{Extract the co-expression list containing the input lncRNA and the co-expressed genes.}
\usage{
getShortPath(lncID, endpoints, corrlist)
}
\arguments{
\item{lncID}{A ensembl id of lncRNA.}

\item{endpoints}{The list of mRNAs co-expressed with this lncRNA.}

\item{corrlist}{The co-expressed correlation list including lncRNA-mRNA and mRNA-mRNA pairs.}
}
\value{
A co-expression list between lncRNAs and target genes, including both lncRNA-mRNA and mRNA-mRNA pairs.
}
\description{
Extract the co-expression list containing the input lncRNA and the co-expressed genes.
}
\examples{
\dontrun{
library(dplyr)
data(CoExpreList_lm_mm)
ID = 'ENSG00000268066'
end = unique((corrlist \%>\% dplyr::filter(simple_type.x == 'lncRNA',
simple_type.y == 'mRNA', row == lncID))$column)
corrlist_shortPath = getShortPath(lncID=ID, endpoints=end,
corrlist=CoExpreList_lm_mm)
}

}
