% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{HeatmapPlot}
\alias{HeatmapPlot}
\title{Draw heatmap plot for units activity score.}
\usage{
HeatmapPlot(object, items, features, mytitle)
}
\arguments{
\item{object}{A scLNC object.}

\item{items}{Attribute for comparison.}

\item{features}{Interesting lncRNA unit name.}

\item{mytitle}{The title of this heatmap.}
}
\value{
A heatmap.
}
\description{
Draw heatmap plot for units activity score.
}
