% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{id_convert_fromgtf}
\alias{id_convert_fromgtf}
\title{Gene names converts to gene ensemble ids.}
\usage{
id_convert_fromgtf(genelist, gtf.info = scLNCgencode)
}
\arguments{
\item{genelist}{A gene names list.}

\item{gtf.info}{Gene annotation from Gencode.}
}
\value{
Return a gene ensemble ids list.
}
\description{
Gene names converts to gene ensemble ids.
}
\examples{
\dontrun{
data(scLNCgencode)
id_convert_fromgtf(genelist=c('LINC00589','LINC00963','LINC01281','CYTOR'),
 gtf.info = scLNCgencode)
}

}
