% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{lnc_network}
\alias{lnc_network}
\title{Differences in functional enrichment between the two groups of units.}
\usage{
lnc_network(GO_file, genelist, lncRNA)
}
\arguments{
\item{GO_file}{Metascape results.}

\item{genelist}{Two mRNA lists from two units.}

\item{lncRNA}{Interesting lncRNA name.}
}
\value{
A network diagram, including differences and same in biological functions and mRNAs.
}
\description{
Differences in functional enrichment between the two groups of units.
}
\examples{
\dontrun{
data(Go2Group_LINC00861)
data(Gene2Group_LINC00861)
lnc_network(GO_file=Go2Group_LINC00861,genelist=Gene2Group_LINC00861)
}

}
