% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{lncRNA_mRNADatabase}
\alias{lncRNA_mRNADatabase}
\title{LncRNA annotation information from lncRNA-mRNA-related databases.}
\usage{
lncRNA_mRNADatabase(features = NULL)
}
\arguments{
\item{features}{Interesting lncRNA ensemble ids, defaults to all lncRNAs.}
}
\value{
Return a statistical graph of lncRNAs from lncRNA-mRNA databases.
}
\description{
The lncRNA-mRNA information are from published databases, including ENCORI, LncRNA2Target v2.0, LncTarD v1.0, MONOCLdb, NPInter v4.0, RISE, RAIN v1.0, RNAInter and lncReg.
}
\examples{
\dontrun{
data(scLNCgencode)
lncRNA_mRNADatabase(features=c('ENSG00000251191','ENSG00000204054',
'ENSG00000235304','ENSG00000222041'))
}

}
