% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{lncRNADatabase}
\alias{lncRNADatabase}
\title{LncRNA annotation information from lncRNA-related databases.}
\usage{
lncRNADatabase(features = NULL)
}
\arguments{
\item{features}{Interesting lncRNA ensemble ids, defaults to all lncRNAs.}
}
\value{
Return a statistical graph of lncRNAs about information related to disease, drug, and function.
}
\description{
lncRNA-related databases included disease, drug, and function related information from Lncbook v1.0, FANTOM5, LncSEA and Lnc2Cancer v3.0 .
}
\examples{
\dontrun{
lncRNADatabase(features=c('ENSG00000251191','ENSG00000204054',
'ENSG00000235304','ENSG00000222041'))
}

}
