% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{load_tdf_enhancer}
\alias{load_tdf_enhancer}
\title{Title The results of TDF enhancer were collated and the regions were mapped to genes.}
\usage{
load_tdf_enhancer(
  PATH = "~/test/packages/_tdf_enhancer_nydgab/enhancer_result.txt"
)
}
\arguments{
\item{PATH}{Input the pathway of rgt-TDF enhancer result.}
}
\value{
The lnRNA-mRNA pairs.
}
\description{
Title The results of TDF enhancer were collated and the regions were mapped to genes.
}
\examples{
\dontrun{
data(CoExpreList)
path_res_enh=run_TDF(pairList=CoExpreList[,1:2], tdf_type = 'enhancer',
ensembl_version = 105)
load_tdf_enhancer(path_res_enh)
}


}
