% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{load_tdf_promoter}
\alias{load_tdf_promoter}
\title{Title The results of TDF promoter were collated and the regions were mapped to genes.}
\usage{
load_tdf_promoter(
  PATH = "~/test/packages/_tdf_promoter_nydgab/promoter_result.txt"
)
}
\arguments{
\item{PATH}{Input the pathway of rgt-TDF promoter result.}
}
\value{
The lnRNA-mRNA pairs.
}
\description{
Title The results of TDF promoter were collated and the regions were mapped to genes.
}
\examples{
\dontrun{
data(CoExpreList)
path_res_pro=run_TDF(pairList=CoExpreList[,1:2], tdf_type = 'promoter',
ensembl_version = 105)
load_tdf_promoter(path_res_pro)
}


}
