% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{lognorm}
\alias{lognorm}
\title{Normalize Data}
\usage{
lognorm(count, scale = 10000)
}
\arguments{
\item{count}{Count matrix.}

\item{scale}{Scale the data. Default is 10000.}
}
\value{
Return a matrix with the normalize and log transformed data.
}
\description{
Normalize the count data in the object.
}
\examples{
\dontrun{
count_filter=matrix(data = rbinom(n = 50, size = 5, prob = 0.2), nrow = 10)
count_norm <- lognorm(count_filter, scale = 10000)
count_norm
}

}
