% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{PairsAnnotation}
\alias{PairsAnnotation}
\title{Annotation co-expressed relationship pairs.}
\usage{
PairsAnnotation(
  object,
  Seq = FALSE,
  Enhancer = FALSE,
  Promoter = FALSE,
  TF = TRUE,
  cytokine = TRUE,
  LMpairs = TRUE,
  direct = FALSE,
  mycorrlist = NULL
)
}
\arguments{
\item{object}{A scLNC object.}

\item{Seq}{Boolean values determining if co-expressed relationship pairs should be annotated the sequence-based pairs information. Default is FALSE.}

\item{Enhancer}{Boolean values determining if co-expressed relationship pairs should be annotated the enhancer information. Default is FALSE.}

\item{Promoter}{Boolean values determining if co-expressed relationship pairs should be annotated the promoter information. Default is FALSE.}

\item{TF}{Boolean values determining if mRNA should be annotated the TF information. Default is TRUE.}

\item{cytokine}{Boolean values determining if mRNA should be annotated the cytokine information. Default is TRUE.}

\item{LMpairs}{Boolean values determining if co-expressed relationship pairs should be annotated the lncRNA-mRNA database information. Default is TRUE.}

\item{direct}{Boolean values determine whether direct or indirect coexpression relation pairs are distinguished. The default value is FALSE.}

\item{mycorrlist}{The co-expressed correlation list including lncRNA-mRNA and mRNA-mRNA pairs.}
}
\value{
Return A scLNC object with annotational pairs.
}
\description{
Annotation co-expressed relationship pairs.
}
\examples{
\dontrun{
data(LNCobject)
scLNC <- PairsAnnotation(object=LNCobject,TF=TRUE,cytokine=TRUE,LMpairs=TRUE)
}

}
