% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{run_LncTar}
\alias{run_LncTar}
\title{Title Run LncTar for lncRNA-mRNA interactions.}
\usage{
run_LncTar(pairList, ensembl_version = 105)
}
\arguments{
\item{pairList}{The lncRNA-mRNA pairs,first column lncRNA, second column mRNA, and the form is ensembl id.}

\item{ensembl_version}{The version of ensemble id. Default is 105.}
}
\value{
A file of lncTar result, including lncRNA, mRNA and the normalized binding free energy (ndG).
}
\description{
Title Run LncTar for lncRNA-mRNA interactions.
}
\examples{
\dontrun{
data(CoExpreList)
path_res_lnctar=run_LncTar(pairList=CoExpreList[,1:2], ensembl_version = 105)
}

}
