% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{run_TDF}
\alias{run_TDF}
\title{Title Run rgt-TDF to identify the target regions related to lncRNA.}
\usage{
run_TDF(pairList, tdf_type = "enhancer", ensembl_version = 105)
}
\arguments{
\item{pairList}{The lncRNA-mRNA pairs,first column lncRNA, second column mRNA, and the form is ensemble id.}

\item{tdf_type}{The target regions include 'enhancer' or 'promoter'.}

\item{ensembl_version}{The version of ensemble id. Default is 105.}
}
\value{
A file of rgt-TDF regiontest result, including DBD binds significantly to the target regions.
}
\description{
Title Run rgt-TDF to identify the target regions related to lncRNA.
}
\examples{
\dontrun{
data(CoExpreList)
path_res_enh=run_TDF(pairList=CoExpreList[,1:2], tdf_type = 'enhancer',
 ensembl_version = 105)
}

}
