% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scCV2}
\alias{scCV2}
\title{Identify highly variable genes.}
\usage{
scCV2(object, limx = c(-2, 2), labelgene = "")
}
\arguments{
\item{object}{A scLNC object.}

\item{limx}{The X-axis range of the fitted curve.}

\item{labelgene}{A vector containing the gene symbols labeled in the plot.}
}
\value{
A curve to the relationship between gene expression and variance of genes.
}
\description{
Identify highly variable genes.
}
