% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_1_CreateLNCobject.R
\name{scLNC_1_CreateLNCobject}
\alias{scLNC_1_CreateLNCobject}
\title{Create a scLNC object for analysis.}
\usage{
scLNC_1_CreateLNCobject(
  count_table1,
  cell_info1,
  geneid = TRUE,
  create = TRUE,
  splitobject = TRUE,
  split.item1
)
}
\arguments{
\item{count_table1}{Gene count matrix with cells as columns and features as rows.}

\item{cell_info1}{A list of cell information. A data.frame where the rows are cell names and the columns are cell informations. Row names in the cell_info1 need to match the column names of the count_table1.}

\item{geneid}{Boolean values determining if gene names convert to Ensemble IDs. Default is TRUE.}

\item{create}{Boolean values determine whether the object is to be created. The default value is TRUE.}

\item{splitobject}{Boolean values determine whether objects should be split into items. Default is TRUE.}

\item{split.item1}{Attribute for splitting.}
}
\value{
Return a scLNC object or a list of scLNC object by different items.
}
\description{
Create a scLNC object for analysis.
}
\examples{
 \dontrun{
data(HCC_rawcount)
data(HCC_cellinfo)

object=scLNC_1_CreateLNCobject (count_table1= HCC_rawcount,
cell_info1= HCC_cellinfo, geneid=TRUE, splitobject=FALSE, split.item1=NA)
object.list=scLNC_1_CreateLNCobject (count_table1= HCC_rawcount,
cell_info1= HCC_cellinfo, geneid=TRUE, splitobject=TRUE,split.item1="Tissue")
}

}
