% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_2_Relation.R
\name{scLNC_2_Relation}
\alias{scLNC_2_Relation}
\title{Perform calculation, filtration and annotation of lncRNA-mRNA co-expression relationship pairs.}
\usage{
scLNC_2_Relation(
  objectInput,
  relation = TRUE,
  relation.fileName,
  pairs.filter = TRUE,
  corcut,
  anno.pairs = TRUE
)
}
\arguments{
\item{objectInput}{A scLNC object including expression matrix slot.}

\item{relation}{Boolean values determine whether to calculate the co-expression of lncRNA-mRNA. The default value is TRUE.}

\item{relation.fileName}{The name of the output file.}

\item{pairs.filter}{The Boolean value determines whether the calculated relationship pairs should be filtered. The default value is TRUE.}

\item{corcut}{Choose the relationship pairs with higher correlation coefficient.}

\item{anno.pairs}{Boolean values determine whether the relationship pairs should be annotated, including TF, cytokine, known lncRNA-mRNA databases, promoter and enhancer information. The default value is TRUE.}
}
\value{
A scLNC object including lncRNA-mRNA co-expression relationship pairs.
}
\description{
Perform calculation, filtration and annotation of lncRNA-mRNA co-expression relationship pairs.
}
\examples{
\dontrun{
data(LNCobject)
LNCobject=scLNC_2_Relation(objectInput=LNCobject,relation.fileName='test',
pairs.filter=TRUE,corcut=0.6,anno.pairs=TRUE)
}

}
