% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_3_Unit.R
\name{scLNC_3_Unit}
\alias{scLNC_3_Unit}
\title{Calculate lncRNA units activity score.}
\usage{
scLNC_3_Unit(
  objectInput,
  AUC = TRUE,
  displayLncRNA = NULL,
  Go = FALSE,
  MetascapeGofile
)
}
\arguments{
\item{objectInput}{A scLNC object including expression matrix slot.}

\item{AUC}{The Boolean value determines whether to calculate the activity of the unit. The default value is TRUE.}

\item{displayLncRNA}{lncRNA list. Only keep a subset of lncRNA units, defaults to all lncRNA units.}

\item{Go}{The Boolean value determines whether to visualize the functional enrichment results of several units. The default value is FALSE.}

\item{MetascapeGofile}{Metascape results.}
}
\value{
Return a scLNC object with AUC slot.
}
\description{
Calculate lncRNA units activity score.
}
\examples{
\dontrun{
data(LNCobject)
LNCobject=scLNC_3_Unit (objectInput=LNCobject,AUC=TRUE,displayLncRNA=NULL)
}

}
