% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_4_Compare.R
\name{scLNC_4_Compare}
\alias{scLNC_4_Compare}
\title{Compare units from different items.}
\usage{
scLNC_4_Compare(
  DEtarget = TRUE,
  ob.ls,
  lncRNA,
  DeGO = TRUE,
  CopareGOfile,
  CopareGeneList
)
}
\arguments{
\item{DEtarget}{Boolean values determine whether to compare composition of units between two groups. The default value is TRUE.}

\item{ob.ls}{A list of objects including two scLNC objects that need to be compared.}

\item{lncRNA}{Interesting lncRNA unit name.}

\item{DeGO}{Boolean values determine whether to compare differences in functional enrichment between the two groups of units. The default value is TRUE.}

\item{CopareGOfile}{A GO_file Metascape results of two groups.}

\item{CopareGeneList}{A data frame including two mRNA lists from two groups.}
}
\value{
Corresponding plot.
}
\description{
Compare units from different items.
}
