% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{splitLRAT}
\alias{splitLRAT}
\title{Splits object into a list of subsetted objects.}
\usage{
splitLRAT(
  object,
  split.item,
  mincell.peritem = 15,
  min.cells.pergene = 0,
  min.gene.percell = 0
)
}
\arguments{
\item{object}{A scLNC object.}

\item{split.item}{Attribute for splitting.}

\item{mincell.peritem}{Include attribute where at least this many cells.}

\item{min.cells.pergene}{Include features detected in at least this many cells. Default is 0.}

\item{min.gene.percell}{Include cells where at least this many genes are detected. Default is 0.}
}
\value{
Return a list of scLNC object, each containing a subset of cells from the original object.
}
\description{
Splits object into a list of subsetted objects.
}
\examples{
\dontrun{
data(HCC_rawcount)
data(HCC_cellinfo)
LNCobject <- createLRAT(count_table = HCC_rawcount, cell_info = HCC_cellinfo,
 min.cells = 15, min.genes = 200, proj.name = "HCC")
LNCobject.list <- splitLRAT(LNCobject, split.item="Tissue")
}

}
