% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{sta_target33}
\alias{sta_target33}
\title{Statistics on the percentage of Enhancer, Promoter and sequence-based.}
\usage{
sta_target33(CoexpPairs)
}
\arguments{
\item{CoexpPairs}{}
}
\value{
A barplot.
}
\description{
Statistics on the percentage of Enhancer, Promoter and sequence-based.
}
