% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{StatGeneNum}
\alias{StatGeneNum}
\title{The difference in the number of genes across different cell types in different experimental condition.}
\usage{
StatGeneNum(
  object,
  genetype = "lncRNA",
  item = "Tissue",
  item.level = c("T", "N", "P"),
  split.by = "majorCluster",
  disorder = c("C04_CD8-LAYN", "C08_CD4-CTLA4", "C10_CD4-CXCL13", "C05_CD8-GZMK",
    "C11_CD4-GNLY", "C09_CD4-GZMA", "C07_CD4-FOXP3", "C03_CD8-SLC4A10", "C01_CD8-LEF1",
    "C06_CD4-CCR7", "C02_CD8-CX3CR1")
)
}
\arguments{
\item{object}{A scLNC object.}

\item{genetype}{'lncRNA' or 'mRNA'.}

\item{item}{A experimental condition variables to group cells by.}

\item{item.level}{The order of experimental condition variables to display.}

\item{split.by}{A variable to split the violin plots by.}

\item{disorder}{The order of cell types to display.}
}
\value{
A violin plot.
}
\description{
The difference in the number of genes across different cell types in different experimental condition.
}
