% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_0_genetic.R
\name{stats_indirect_pairs}
\alias{stats_indirect_pairs}
\title{Identify the indirect and direct correlated mRNAs with lncRNAs.}
\usage{
stats_indirect_pairs(corrlist)
}
\arguments{
\item{corrlist}{The co-expressed correlation list including lncRNA-mRNA and mRNA-mRNA pairs.}
}
\value{
List of lncRNA-mRNA pairs with indirect and direct correlation and statistical diagram.
}
\description{
Identify the indirect and direct correlated mRNAs with lncRNAs.
}
\examples{
\dontrun{
data(CoExpreList_lm_mm)
direct.list=stats_indirect_pairs(corrlist=CoExpreList_lm_mm)
}

}
