% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLNC_3_Unit.R
\name{scLNC_3_Unit}
\alias{scLNC_3_Unit}
\title{Calculate lncRNA units activity score.}
\usage{
scLNC_3_Unit(
  objectInput,
  AUC = TRUE,
  displayLncRNA = NULL,
  DEAUC = TRUE,
  item.add = NULL,
  DEitem = "majorCluster"
)
}
\arguments{
\item{objectInput}{A scLNC object including expression matrix slot.}

\item{AUC}{The Boolean value determines whether to calculate the activity of the unit. The default value is TRUE.}

\item{displayLncRNA}{lncRNA list. Only keep a subset of lncRNA units, defaults to all lncRNA units.}

\item{DEAUC}{Boolean values determine whether to compare differential activity of units between two groups. The default value is TRUE.}

\item{item.add}{A new list of cell information, including new cell item column.}

\item{DEitem}{ttribute for comparison.}
}
\value{
Return a scLNC object with AUC slot.
}
\description{
Calculate lncRNA units activity score.
}
\examples{
\dontrun{
data(LNCobject)
LNCobject=scLNC_3_Unit (objectInput=LNCobject,AUC=TRUE,displayLncRNA=NULL,
DEAUC=TRUE,item.add=NULL,DEitem='majorCluster')
}

}
